refactor to separate concerns
